backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "json-mysql-bulk"
version = "1.0.0"
description = "Sistema profissional para carga em massa de arquivos JSON em bancos MySQL"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Pedro Mendes", email = "contato@example.com"}
]

requires-python = ">=3.10"

dependencies = [
    "pandas>=2.0.0",
    "SQLAlchemy>=2.0.0",
    "PyMySQL>=1.1.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
perf = [
    "ujson>=5.8.0",
]

[project.urls]
Repository = "https://github.com/usuario/json-mysql-bulk"
Issues = "https://github.com/usuario/json-mysql-bulk/issues"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
markers = [
    "e2e: marca teste como integração com MySQL",
]
testpaths = ["tests"]
